requirements.txt
Flask==2.3.2
pandas==2.1.1
numpy==1.25.2
Flask-PyMongo==2.3.0
google-generativeai==0.2.0
scikit-learn==1.3.1
Werkzeug==2.3.6
bson==0.5.10
